Full JWT verifcation added

+++ src/main/java/com/example/MS_Backend/controllers/MovieController.java
@@ -1,11 +1,15 @@
 package com.example.MS_Backend.controllers;
 
 import com.example.MS_Backend.models.User;
+import com.example.MS_Backend.services.JWTService;
 import com.example.MS_Backend.services.MovieService;
 import com.fasterxml.jackson.databind.JsonNode;
+import io.jsonwebtoken.*;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseCookie;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -17,33 +21,87 @@ public class MovieController {
     @Autowired
     private MovieService service;
 
-    @GetMapping("/genres")
-    public JsonNode findGenres() throws Exception {
-        return service.getGenres();
+    @Autowired
+    private JWTService jwtService;
+
+    @GetMapping("/genres/{username}")
+    public JsonNode findGenres(@PathVariable String username, HttpServletRequest request, HttpServletResponse response) throws Exception {
+        if(verify(request, response, username)) {
+            return service.getGenres();
+        }
+        throw new RuntimeException("Authentication error");
     }
 
-    @GetMapping("/mymovies/{filters}")
-    public List<Map<String, String>> getMovies(@PathVariable String filters) throws Exception {
-        return service.searchMovies(filters);
+    @GetMapping("/mymovies/{username}/{filters}")
+    public List<Map<String, String>> getMovies(@PathVariable String username, @PathVariable String filters, HttpServletRequest request, HttpServletResponse response) throws Exception {
+        if(verify(request, response, username)) {
+            return service.searchMovies(filters);
+        }
+        throw new RuntimeException("Authentication error");
     }
 
     @PutMapping("/addmovies/{username}")
-    public void addMovies(@PathVariable String username, @RequestBody List<Map<String, String>> movies) {
-        service.addMovies(username, movies);
+    public void addMovies(@PathVariable String username, @RequestBody List<Map<String, String>> movies, HttpServletRequest request, HttpServletResponse response) {
+        if(verify(request, response, username)) {
+            service.addMovies(username, movies);
+        }
     }
 
     @GetMapping("/getmovies/{username}")
-    public JsonNode getMoviesByUser(@PathVariable String username, HttpServletRequest request) throws Exception {
-        Cookie[] cookies = request.getCookies();
-        for(Cookie cookie : cookies) {
-            System.out.println("This cookie is: " + cookie.getName() + ": " + cookie.getValue());
+    public JsonNode getMoviesByUser(@PathVariable String username, HttpServletRequest request, HttpServletResponse response) throws Exception {
+        if(verify(request, response, username)) {
+            return service.getMovies(username);
         }
-        return service.getMovies(username);
+        throw new RuntimeException("Authentication error");
     }
 
     @PatchMapping("/deletemovie/{username}/{title}")
-    public void removeMovie(@PathVariable String username, @PathVariable String title) {
-        service.deleteMovies(username, title);
+    public void removeMovie(@PathVariable String username, @PathVariable String title, HttpServletRequest request, HttpServletResponse response) {
+        if(verify(request, response, username)) {
+            service.deleteMovies(username, title);
+        }
+    }
+
+    public boolean verify(HttpServletRequest request, HttpServletResponse response, String username) {
+        Cookie[] cookies = request.getCookies();
+        boolean expired = false;
+
+        if(cookies != null) {
+            for(Cookie cookie : cookies) {
+                String token = cookie.getValue();
+                try {
+                    Claims claims = Jwts.parser()
+                            .verifyWith(jwtService.getKey())
+                            .build()
+                            .parseSignedClaims(token)
+                            .getPayload();
+
+                } catch(ExpiredJwtException e) {
+                    expired = true;
+                } catch(JwtException e) {
+                    System.out.println("This token is not valid");
+                    throw new RuntimeException(e);
+                }
+            }
+        } else {
+            System.out.println("Cookie expired or removed");
+            expired = true;
+        }
+
+        if(expired) {
+            System.out.println("Token or cookies expired");
+            String token = jwtService.generateToken(username);
+
+            ResponseCookie cookie = ResponseCookie.from("token", token)
+                    .httpOnly(true)
+                    .secure(true)
+                    .path("/")
+                    .maxAge(30)
+                    .sameSite("None")
+                    .build();
+            response.setHeader("Set-Cookie", cookie.toString());
+        }
+        return true;
     }
 
 }
